Update ForCE.m
This update is pretty major so be careful when updating from the previous version. There are some trivial corrections to the sea level rise component of the code after completing some numerical testing. The more major bits are attempts to give better convergence to the KF. This sets b1=1. a2 is set to zero. Finally, k2 is set equal to k1. A future version is in prep. which removes redundant variables from the state matrix.